+++ src/main/kotlin/com/github/alaanor/candid/editing/CandidCommenter.kt
@@ -1,0 +1,11 @@
+package com.github.alaanor.candid.editing
+
+import com.intellij.lang.Commenter
+
+class CandidCommenter : Commenter {
+    override fun getLineCommentPrefix(): String = "//"
+    override fun getBlockCommentPrefix(): String = "/*"
+    override fun getBlockCommentSuffix(): String = "*/"
+    override fun getCommentedBlockCommentPrefix(): String? = null
+    override fun getCommentedBlockCommentSuffix(): String? = null
+}